requirements.txt
google-genai
chromadb
sentence-transformers
streamlit